authentication done but still need to maintain the business logic
